requirements.txt
requests == 2.24.0
pyserial == 3.4
pydrive == 1.3.1
opencv-python == 4.4.0.44
pydrive == 1.3.1
cvlib == 0.2.5
python-dotenv == 0.15.0
tensorflow == 2.3.1
